% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_dlink.R
\name{calcular_dlink}
\alias{calcular_dlink}
\title{Calcula o valor D-LINK (ordem do trecho jusante)}
\usage{
calcular_dlink(df, col_id, col_jus, col_ordem)
}
\arguments{
\item{df}{Data frame contendo os trechos da rede hidrografica}

\item{col_id}{Nome da coluna com ID unico dos trechos (ex: "cotrecho")}

\item{col_jus}{Nome da coluna com o ID do trecho jusante (ex: "nutrjus")}

\item{col_ordem}{Nome da coluna com a ordem do trecho (ex: "link_mag", "ordem", etc.)}
}
\value{
O mesmo \code{df}, com uma nova coluna \code{D_LINK}
}
\description{
Esta funcao adiciona a uma tabela de trechos fluviais a coluna \code{D_LINK}, que representa
o valor da ordem (ex: Link Magnitude) do trecho jusante.
}
\examples{
df <- data.frame(
  id = c(1, 2, 3),
  jus = c(2, 3, NA),
  ordem = c(1, 2, 3)
)
calcular_dlink(df, col_id = "id", col_jus = "jus", col_ordem = "ordem")
}
