% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network_graph.R
\name{create_network_graph}
\alias{create_network_graph}
\title{Create a directed graph from a stream network}
\usage{
create_network_graph(network, col_id, col_downstream)
}
\arguments{
\item{network}{\code{sf} or \code{SpatVector} object representing the stream network reaches}

\item{col_id}{Name of the column with the unique ID of each reach}

\item{col_downstream}{Name of the column with the downstream reach ID}
}
\value{
Directed \code{igraph} object representing the stream network
}
\description{
Builds an \code{igraph} graph from an \code{sf} or \code{SpatVector} object
containing stream reaches. Requires columns with unique reach IDs and their
respective downstream reach IDs.
}
\examples{
library(igraph)
library(sf)
lines <- st_sfc(
  st_linestring(matrix(c(0,0, 1,1), ncol = 2, byrow = TRUE)),
  st_linestring(matrix(c(1,1, 2,2), ncol = 2, byrow = TRUE)),
  st_linestring(matrix(c(2,2, 3,3), ncol = 2, byrow = TRUE))
)
data <- data.frame(id = c(1, 2, 3), downstream = c(2, 3, NA))
network <- st_sf(data, geometry = lines)
g <- create_network_graph(network, col_id = "id", col_downstream = "downstream")
plot(g)
}
