% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criar_grafo_rede.R
\name{criar_grafo_rede}
\alias{criar_grafo_rede}
\title{Cria grafo direcionado a partir de rede fluvial}
\usage{
criar_grafo_rede(rede, col_id, col_jus)
}
\arguments{
\item{rede}{Objeto \code{sf} ou \code{SpatVector} representando os trechos da rede hidrográfica}

\item{col_id}{Nome da coluna com o ID único de cada trecho}

\item{col_jus}{Nome da coluna com o ID do trecho jusante}
}
\value{
Objeto \code{igraph} direcionado representando a rede hidrográfica
}
\description{
Esta função constrói um grafo \code{igraph} a partir de um objeto \code{sf} ou \code{SpatVector} contendo trechos fluviais.
É necessário que existam colunas com IDs únicos dos trechos e seus respectivos trechos jusantes.
}
\examples{
library(igraph)
library(sf)
linhas <- st_sfc(
  st_linestring(matrix(c(0,0, 1,1), ncol = 2, byrow = TRUE)),
  st_linestring(matrix(c(1,1, 2,2), ncol = 2, byrow = TRUE)),
  st_linestring(matrix(c(2,2, 3,3), ncol = 2, byrow = TRUE))
)
dados <- data.frame(id = c(1, 2, 3), jus = c(2, 3, NA))
rede <- st_sf(dados, geometry = linhas)
g <- criar_grafo_rede(rede, col_id = "id", col_jus = "jus")
plot(g)
}
