% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_distancia_foz.R
\name{calcular_distancia_foz}
\alias{calcular_distancia_foz}
\title{Calcula a distância acumulada até a foz}
\usage{
calcular_distancia_foz(rede, foz, peso = "length_km")
}
\arguments{
\item{rede}{Objeto \code{igraph} representando a rede hidrográfica (com pesos nas arestas)}

\item{foz}{Vetor de IDs dos nós que representam a foz}

\item{peso}{Nome do atributo de peso nas arestas (ex: 'length_km')}
}
\value{
Vetor nomeado com distâncias acumuladas até a foz
}
\description{
Esta função calcula a distância de cada trecho fluvial até a foz usando grafos direcionais.
}
