% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_link_magnitude.R
\name{calcular_link_magnitude}
\alias{calcular_link_magnitude}
\title{Calcula Link Magnitude (Shreve) de uma rede hidrográfica}
\usage{
calcular_link_magnitude(grafo)
}
\arguments{
\item{grafo}{Objeto \code{igraph} direcionado representando a rede hidrográfica}
}
\value{
Vetor nomeado com valores inteiros de link magnitude para cada nó
}
\description{
Esta função calcula o Link Magnitude (ordem de Shreve) para cada trecho da rede fluvial,
assumindo que a rede é um grafo acíclico direcionado (DAG).
}
\examples{
g <- igraph::make_graph(edges = c("1", "2", "2", "3", "3", "4"), directed = TRUE)
calcular_link_magnitude(g)
}
