% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGDrivers-package.R
\docType{package}
\name{RGDrivers-package}
\alias{RGDrivers}
\alias{RGDrivers-package}
\title{RGDrivers: Analysis of Stream Network Topology and Order}
\description{
Provides tools for analyzing stream networks, including graph construction, calculation of Link Magnitude, D-LINK, and export of spatial data. The Link Magnitude (Shreve stream order) method follows Shreve (1966) \doi{10.1086/627137}. The D-LINK metric follows Osborne & Wiley (1992) \doi{10.1139/f92-076}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/diogosbr/RGDrivers}
  \item Report bugs at \url{https://github.com/diogosbr/RGDrivers/issues}
}

}
\author{
\strong{Maintainer}: Diogo S. B. Rocha \email{diogosbr@gmail.com}

Authors:
\itemize{
  \item Diogo S. B. Rocha \email{diogosbr@gmail.com}
  \item Renata T. Capellão \email{renatacapellao@gmail.com}
}

}
\keyword{internal}
