% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGDrivers-package.R
\docType{package}
\name{RGDrivers-package}
\alias{RGDrivers}
\alias{RGDrivers-package}
\title{RGDrivers: Tools for Stream Network Analysis}
\description{
Provides tools for analyzing stream networks, including graph construction, calculation of Link Magnitude (Shreve stream order), D-LINK, and export of spatial data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/diogosbr/RGDrivers}
  \item Report bugs at \url{https://github.com/diogosbr/RGDrivers/issues}
}

}
\author{
\strong{Maintainer}: Diogo Rocha \email{diogo@gmail.com}

Authors:
\itemize{
  \item Diogo Rocha \email{diogo@gmail.com}
  \item Renata Capellão \email{renatacapellao@gmail.com}
  \item Luiz Oliveira \email{gustavo.if.ufrj@gmail.com}
}

}
\keyword{internal}
