% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGDrivers-package.R
\docType{package}
\name{RGDrivers-package}
\alias{RGDrivers}
\alias{RGDrivers-package}
\title{RGDrivers: Ferramentas para Análise de Redes Hidrográficas}
\description{
Um pacote para análise de redes hidrográficas, incluindo construção de grafos, cálculo de Link Magnitude (Shreve), D-LINK e exportação de dados espaciais.
}
\author{
\strong{Maintainer}: Diogo Rocha \email{diogo@gmail.com}

Authors:
\itemize{
  \item Renata Capellão \email{renatacapellao@gmail.com}
  \item Luiz Oliveira \email{gustavo.if.ufrj@gmail.com}
}

}
\keyword{internal}
